+++ nodes/Payi/descriptions/index.ts
@@ -1,3 +1,4 @@
 export { providerFields } from './providerFields';
 export { trackingFields } from './trackingFields';
 export { outputFields } from './outputFields';
+export { chatModelFields } from './chatModelFields';